driver: Add support for RancherVM
Add a parameter called "--vm" to specify this volume is created for Rancher VM.
The behavior would be:

1. For directory based driver(e.g. vfs, glusterfs), it would create a image
file(if not exists), then make Rancher VM use it as a disk image.

2. For device based driver(e.g. device mapper, EBS), we can do an optimization,
to pass the device directly to QEmu inside container as a device, which would be
greatly benefit from the performance.

We would implement the first kind of behavior first.