Add option to create volume during docker mount call
This fix works around a bug in docker that prevents you from either
creating or deleting a volume if its been deleted via the convoy api
before being deleted via docker.

The issue arises when you try to recreate a volume that has been
deleted in the above manner. Docker will report it has been created
because it thinks it exists, but the volume will not show up in `docker
volume ls` or `docker volume inspect` and `docker volume rm`s will fail
without ever calling the driver to remvoe the volume.

So this change gives us the ability to create a volume when
docker.mount is called. This will bring the two systems back in sync
and allow the mount to happen.