Time outs and Force Detach tagging
This update now adds timeouts to waitForAttaching and waitForDetaching functions in convoy.
Upon a time out force detaches a volume from the instance and sends an err to the daemon
which will cause a reschedule on Mesos.
If a timeout (and eventually a force detach) is triggered, then the infrastructure is
marked with ForceDetached tag set to true. Monitoring can handle it from there on.
This release also fixes code to get clustername tag that volumes get to happen the right way.