Update state if it has volume that's not attached
Before this the volume mount would fail when convoy thinks there is
a volume json file in the driver state but it was probably detached
by bypassing convoy in the past, thus not updating the state.

Now it updates the state by deleting the volume.

Also use `aws.StringValue(s *string)` since just dereferencing a
pointer without it can cause a segmentation fault.